Refactor: Outsource state extraction into envs/
