+++ envs/asterix/valuation.py
@@ -1,4 +1,4 @@
-from nsfr.utils import bool_to_probs
+from nsfr.utils.common import bool_to_probs
 
 
 def obj_type(z, a):
