+++ nsfr/nsfr/utils/common.py
