+++ nsfr/nsfr/utils/logic.py
@@ -1,8 +1,8 @@
-from .infer import InferModule, ClauseInferModule, ClauseBodyInferModule
-from .tensor_encoder import TensorEncoder
-from .fol.logic import *
-from .fol.data_utils import DataUtils
-from .fol.language import DataType
+from nsfr.infer import InferModule, ClauseInferModule, ClauseBodyInferModule
+from nsfr.tensor_encoder import TensorEncoder
+from nsfr.fol.logic import *
+from nsfr.fol.data_utils import DataUtils
+from nsfr.fol.language import DataType
 
 p_ = Predicate('.', 1, [DataType('spec')])
 false = Atom(p_, [Const('__F__', dtype=DataType('spec'))])
