+++ nudge/agents/utils_atari.py
@@ -1,118 +1,7 @@
 from enum import Enum
-from typing import Final, List
 
 import numpy as np
 import torch
-
-
-def extract_logic_state_atari(state, args, noise=False, device=None):
-    if 'freeway' in args.env.lower():
-        num_of_feature = 6
-        num_of_object = 11
-        representation = state
-        extracted_states = np.zeros((num_of_object, num_of_feature))
-        for i, entity in enumerate(representation):
-            if entity.category == "Chicken" and i == 0:
-                extracted_states[0][0] = 1
-                extracted_states[0][-2:] = entity.xy
-            elif entity.category == 'Car':
-                extracted_states[i-1][1] = 1
-                extracted_states[i-1][-2:] = entity.xy
-    elif 'asterix' in args.env.lower():
-        num_of_feature = 6
-        num_of_object = 11
-        representation = state
-        extracted_states = np.zeros((num_of_object, num_of_feature))
-        for i, entity in enumerate(representation):
-            if entity.category == "Player":
-                extracted_states[i][0] = 1
-                extracted_states[i][-2:] = entity.xy
-            elif entity.category == 'Enemy':
-                extracted_states[i][1] = 1
-                extracted_states[i][-2:] = entity.xy
-            elif "Reward" in entity.category:
-                extracted_states[i][2] = 1
-                extracted_states[i][-2:] = entity.xy
-            else:
-                extracted_states[i][3] = 1
-                extracted_states[i][-2:] = entity.xy
-    else:
-        raise NotImplementedError
-
-    def simulate_prob(extracted_states, num_of_objs, key_picked):
-        for i, obj in enumerate(extracted_states):
-            obj = add_noise(obj, i, num_of_objs)
-            extracted_states[i] = obj
-        if key_picked:
-            extracted_states[:, 1] = 0
-        return extracted_states
-
-    def add_noise(obj, index_obj, num_of_objs):
-        mean = torch.tensor(0.2)
-        std = torch.tensor(0.05)
-        noise = torch.abs(torch.normal(mean=mean, std=std)).item()
-        rand_noises = torch.randint(1, 5, (num_of_objs - 1,)).tolist()
-        rand_noises = [i * noise / sum(rand_noises) for i in rand_noises]
-        rand_noises.insert(index_obj, 1 - noise)
-
-        for i, noise in enumerate(rand_noises):
-            obj[i] = rand_noises[i]
-        return obj
-
-    # if noise:
-    #     extracted_states = simulate_prob(extracted_states, num_of_object, key_picked)
-    states = torch.tensor(np.array(extracted_states), dtype=torch.float32, device=device).unsqueeze(0)
-    return states
-
-
-def extract_neural_state_atari(state, args, device=None):
-    if 'freeway' in args.env.lower():
-        raw_state = []
-        for i, inst in enumerate(state):
-            if inst.category == "Chicken" and i == 1:
-                raw_state.append([1, 0, 0, 0] + list(inst.xy))
-            elif inst.category == "Car":
-                raw_state.append([0, 1, 0, 0] + list(inst.xy))
-    elif 'asterix' in args.env.lower():
-        raw_state = []
-        for i, inst in enumerate(state):
-            if inst.category == "Player" and i == 0:
-                raw_state.append([1, 0, 0, 0] + list(inst.xy))
-            elif inst.category == "Enemy":
-                raw_state.append([0, 1, 0, 0] + list(inst.xy))
-            elif "Reward" in inst.category:
-                raw_state.append([0, 0, 1, 0] + list(inst.xy))
-            else:
-                raw_state.append([0, 0, 0, 1] + list(inst.xy))
-            # elif inst.category == "Cauldron":
-            #     raw_state.append([0, 1] + list(inst.xy))
-            # elif inst.category == "Helmet":
-            #     raw_state.append([0, 1] + list(inst.xy))
-            # elif inst.category == "Shield":
-            #     raw_state.append([0, 1] + list(inst.xy))
-            # elif inst.category == "Lamp":
-            #     raw_state.append([0, 1] + list(inst.xy))
-            # elif inst.category == "Apple":
-            #     raw_state.append([0, 1] + list(inst.xy))
-            # elif inst.category == "Fish":
-            #     raw_state.append([0, 1] + list(inst.xy))
-            # elif inst.category == "Meat":
-            #     raw_state.append([0, 1] + list(inst.xy))
-        if len(raw_state) < 11:
-            raw_state.extend([[0] * 6 for _ in range(11 - len(raw_state))])
-        
-    else:
-        print("Not yet implemented, utils_atari l 64")
-        exit(1)
-    state = np.array(raw_state).reshape(-1)
-    return torch.tensor(state).to(device)
-
-# def extract_neural_state_atari(atari, args):
-#     model_input = sample_to_model_input((extract_state(atari), []))
-#     model_input = collate([model_input])
-#     state = model_input['state']
-#     state = torch.cat([state['base'], state['entities']], dim=1)
-#     return state
 
 
 def for_each_tensor(o, fn):
