+++ nudge/agents/utils_getout.py
@@ -1,160 +1,3 @@
-from enum import Enum
-from typing import List, Final
-import numpy as np
-import torch
-
-
-def extract_logic_state_getout(coin_jump, args, noise=False):
-    if args.env == 'getoutplus':
-        num_of_feature = 6
-        num_of_object = 8
-        representation = coin_jump.level.get_representation()
-        # import ipdb; ipdb.set_trace()
-        extracted_states = np.zeros((num_of_object, num_of_feature))
-        for entity in representation["entities"]:
-            if entity[0].name == 'PLAYER':
-                extracted_states[0][0] = 1
-                extracted_states[0][-2:] = entity[1:3]
-                # 27 is the width of map, this is normalization
-                # extracted_states[0][-2:] /= 27
-            elif entity[0].name == 'KEY':
-                extracted_states[1][1] = 1
-                extracted_states[1][-2:] = entity[1:3]
-                # extracted_states[1][-2:] /= 27
-            elif entity[0].name == 'DOOR':
-                extracted_states[2][2] = 1
-                extracted_states[2][-2:] = entity[1:3]
-                # extracted_states[2][-2:] /= 27
-            elif entity[0].name == 'GROUND_ENEMY':
-                extracted_states[3][3] = 1
-                extracted_states[3][-2:] = entity[1:3]
-                # extracted_states[3][-2:] /= 27
-            elif entity[0].name == 'GROUND_ENEMY2':
-                extracted_states[4][3] = 1
-                extracted_states[4][-2:] = entity[1:3]
-                # extracted_states[3][-2:] /= 27
-            elif entity[0].name == 'GROUND_ENEMY3':
-                extracted_states[5][3] = 1
-                extracted_states[5][-2:] = entity[1:3]
-            elif entity[0].name == 'BUZZSAW1':
-                extracted_states[6][3] = 1
-                extracted_states[6][-2:] = entity[1:3]
-            elif entity[0].name == 'BUZZSAW2':
-                extracted_states[7][3] = 1
-                extracted_states[7][-2:] = entity[1:3]
-    else:
-        """
-        extract state to metric
-        input: coin_jump instance
-        output: extracted_state to be explained
-        set noise to True to add noise
-    
-        x:  agent, key, door, enemy, position_X, position_Y
-        y:  obj1(agent), obj2(key), obj3(door)，obj4(enemy)
-    
-        To be changed when using object-detection tech
-        """
-        num_of_feature = 6
-        num_of_object = 4
-        representation = coin_jump.level.get_representation()
-        extracted_states = np.zeros((num_of_object, num_of_feature))
-        for entity in representation["entities"]:
-            if entity[0].name == 'PLAYER':
-                extracted_states[0][0] = 1
-                extracted_states[0][-2:] = entity[1:3]
-                # 27 is the width of map, this is normalization
-                # extracted_states[0][-2:] /= 27
-            elif entity[0].name == 'KEY':
-                extracted_states[1][1] = 1
-                extracted_states[1][-2:] = entity[1:3]
-                # extracted_states[1][-2:] /= 27
-            elif entity[0].name == 'DOOR':
-                extracted_states[2][2] = 1
-                extracted_states[2][-2:] = entity[1:3]
-                # extracted_states[2][-2:] /= 27
-            elif entity[0].name == 'GROUND_ENEMY':
-                extracted_states[3][3] = 1
-                extracted_states[3][-2:] = entity[1:3]
-                # extracted_states[3][-2:] /= 27
-
-    if sum(extracted_states[:, 1]) == 0:
-        key_picked = True
-    else:
-        key_picked = False
-
-    def simulate_prob(extracted_states, num_of_objs, key_picked):
-        for i, obj in enumerate(extracted_states):
-            obj = add_noise(obj, i, num_of_objs)
-            extracted_states[i] = obj
-        if key_picked:
-            extracted_states[:, 1] = 0
-        return extracted_states
-
-    def add_noise(obj, index_obj, num_of_objs):
-        mean = torch.tensor(0.2)
-        std = torch.tensor(0.05)
-        noise = torch.abs(torch.normal(mean=mean, std=std)).item()
-        rand_noises = torch.randint(1, 5, (num_of_objs - 1,)).tolist()
-        rand_noises = [i * noise / sum(rand_noises) for i in rand_noises]
-        rand_noises.insert(index_obj, 1 - noise)
-
-        for i, noise in enumerate(rand_noises):
-            obj[i] = rand_noises[i]
-        return obj
-
-    if noise:
-        extracted_states = simulate_prob(extracted_states, num_of_object, key_picked)
-    states = torch.tensor(np.array(extracted_states), dtype=torch.float32, device="cuda:0").unsqueeze(0)
-    return states
-
-
-def extract_neural_state_getout(getout, args):
-    model_input = sample_to_model_input((extract_state(getout), []))
-    model_input = collate([model_input])
-    state = model_input['state']
-    state = torch.cat([state['base'], state['entities']], dim=1)
-    return state
-
-
-def for_each_tensor(o, fn):
-    if isinstance(o, torch.Tensor):
-        return fn(o)
-    if isinstance(o, list):
-        for i, e in enumerate(o):
-            o[i] = for_each_tensor(e, fn)
-        return o
-    if isinstance(o, dict):
-        for k, v in o.items():
-            o[k] = for_each_tensor(v, fn)
-        return o
-    raise ValueError("unexpected object type")
-
-
-def collate(samples, to_cuda=True, double_to_float=True):
-    samples = torch.utils.data._utils.collate.default_collate(samples)
-    if double_to_float:
-        samples = for_each_tensor(samples, lambda tensor: tensor.float() if tensor.dtype == torch.float64 else tensor)
-    if to_cuda:
-        samples = for_each_tensor(samples, lambda tensor: tensor.cuda())
-    return samples
-
-
-def extract_state(coin_jump):
-    repr = coin_jump.level.get_representation()
-    repr["reward"] = coin_jump.level.reward
-    repr["score"] = coin_jump.score
-
-    for entity in repr["entities"]:
-        # replace all enums (e.g. EntityID) with int values
-        for i, v in enumerate(entity):
-            if isinstance(v, Enum):
-                entity[i] = v.value
-            if isinstance(v, bool):
-                entity[i] = int(v)
-
-    return repr
-
-
 ENTITY_ENCODING_LENGTH = 9
 PLAYER_IDX = 0 * ENTITY_ENCODING_LENGTH
 FLAG_IDX = 1 * ENTITY_ENCODING_LENGTH
@@ -219,47 +62,6 @@ def fixed_size_entity_representation(state, swap_coins=None):
         tr_entities[COIN1_IDX:COIN1_IDX + ENTITY_ENCODING_LENGTH] = temp_coin
 
     return tr_entities, swap_coins
-
-
-def sample_to_model_input(sample, no_dict=False, include_score=False):
-    """
-    :param sample:  tuple: (representation (use extract_state), explicit_action (use unify_coin_jump_actions))
-    :return: {state: {base:[...], entities:[...]}, action:0..9}
-    """
-    state = sample[0]
-    action = sample[1]
-
-    # tr_entities = replace_bools(fixed_size_entity_representation(state))
-    tr_entity, swap_coins = fixed_size_entity_representation(state)
-    tr_entities, swap_coins = replace_bools(tr_entity, swap_coins)
-    if no_dict:
-        # ignores the action and returns a single [60] array
-        return [
-            0,
-            0,
-            state['level']['reward_key'],
-            state['level']['reward_powerup'],
-            state['level']['reward_enemy'],
-            state['score'] if include_score else 0,
-            *tr_entities
-        ]
-
-    tr_state = {
-        'base': np.array([
-            0,
-            0,
-            state['level']['reward_key'],
-            state['level']['reward_powerup'],
-            state['level']['reward_enemy'],
-            state['score'] if include_score else 0,
-        ]),
-        'entities': np.array(tr_entities)
-    }
-
-    return {
-        "state": tr_state,
-        "action": action
-    }
 
 
 def distance(er, e0, e1):
