+++ nudge/agents/utils_loot.py
@@ -4,137 +4,6 @@ import torch
 device = torch.device('cuda:0')
 
 
-def extract_logic_state_loot(state, args):
-    """
-    [X,Y]
-    [agent,key_b,door_b,key_g,door_g,key_r,door_r]
-    """
-    states = torch.from_numpy(state['positions']).squeeze()
-    if args.env == 'lootplus':
-        # input shape: [X,Y]* [agent,key_b,door_b,key_g,door_g,key_r,door_r]
-        # output shape:[agent, key, door, blue, green, red ,got_key, X, Y]
-        extracted_state = torch.tensor([
-            [1, 0, 0, 0, 0, 0, 0, 0, 0],
-            [0, 1, 0, 1, 0, 0, 0, 0, 0],
-            [0, 0, 1, 1, 0, 0, 0, 0, 0],
-            [0, 1, 0, 0, 1, 0, 0, 0, 0],
-            [0, 0, 1, 0, 1, 0, 0, 0, 0],
-            [0, 1, 0, 0, 0, 1, 0, 0, 0],
-            [0, 0, 1, 0, 0, 1, 0, 0, 0]], dtype=torch.float64)
-        extracted_state[:, -2:] = states[:]
-        for i, state in enumerate(extracted_state):
-            if state[-1] == 0 and state[-2] == 0:
-                extracted_state[i] = torch.zeros((1, 9))
-            elif i in [2, 4, 6] and state[-1] != 0 and extracted_state[i - 1][1] == 0:
-                extracted_state[i][-3] = 1
-    elif args.env == 'loothard':
-        # input shape: [X,Y]* [agent, key_b, door_b, key_g, door_g, exit]
-        # output shape:[agent, key, door, blue, green, exit,got_key, X, Y]
-        extracted_state = torch.tensor([
-            [1, 0, 0, 0, 0, 0, 0, 0, 0],
-            [0, 1, 0, 1, 0, 0, 0, 0, 0],
-            [0, 0, 1, 1, 0, 0, 0, 0, 0],
-            [0, 1, 0, 0, 1, 0, 0, 0, 0],
-            [0, 0, 1, 0, 1, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0, 1, 0, 0, 0]], dtype=torch.float64)
-        states = states[0:6]
-        extracted_state[:, -2:] = states[:]
-        for i, state in enumerate(extracted_state):
-            if state[-1] == 0 and state[-2] == 0:
-                extracted_state[i] = torch.zeros((1, 9))
-            elif i in [2, 4] and state[-1] != 0 and extracted_state[i - 1][1] == 0: # setting got_key
-                extracted_state[i][-3] = 1
-
-    elif args.env == "loot":
-        # input shape: [X,Y]* [agent,key_b,door_b,key_g,door_g]
-        # output shape:[agent, key, door, blue, red ,got_key, X, Y]
-        extracted_state = torch.tensor([
-            [1, 0, 0, 0, 0, 0, 0, 0],
-            [0, 1, 0, 1, 0, 0, 0, 0],
-            [0, 0, 1, 1, 0, 0, 0, 0],
-            [0, 1, 0, 0, 1, 0, 0, 0],
-            [0, 0, 1, 0, 1, 0, 0, 0]], dtype=torch.float64)
-        states = states[0:5]
-        extracted_state[:, -2:] = states[:]
-        for i, state in enumerate(extracted_state):
-            # 0 mean object does not exist
-            if state[-1] == 0 and state[-2] == 0:
-                # then set to all attributes 0
-                extracted_state[i] = torch.zeros((1, 8))
-            # if key = 0 but door !=0, means key of this door has picked
-            elif i in [2, 4] and state[-1] != 0 and extracted_state[i - 1][1] == 0:
-                extracted_state[i][-3] = 1
-    elif args.env == "lootcolor":
-        # input shape: [X,Y]* [agent,key_b,door_b,key_g,door_g]
-        # output shape:[agent, key, door, green, brown ,got_key, X, Y]
-        extracted_state = torch.tensor([
-            [1, 0, 0, 0, 0, 0, 0, 0],
-            [0, 1, 0, 1, 0, 0, 0, 0],
-            [0, 0, 1, 1, 0, 0, 0, 0],
-            [0, 1, 0, 0, 1, 0, 0, 0],
-            [0, 0, 1, 0, 1, 0, 0, 0]], dtype=torch.float64)
-        states = states[0:5]
-        extracted_state[:, -2:] = states[:]
-        for i, state in enumerate(extracted_state):
-            # 0 mean object does not exist
-            if state[-1] == 0 and state[-2] == 0:
-                # then set to all attributes 0
-                extracted_state[i] = torch.zeros((1, 8))
-            # if key = 0 but door !=0, means key of this door has picked
-            elif i in [2, 4] and state[-1] != 0 and extracted_state[i - 1][1] == 0:
-                extracted_state[i][-3] = 1
-    extracted_state = extracted_state.unsqueeze(0)
-    return extracted_state.to(device)
-
-
-def extract_neural_state_loot(state, args):
-    state = state['positions']
-
-    if args.env == 'lootplus':
-        raw_state = np.array([[0, 0, 0, 0],
-                              [0, 0, 1, 1],
-                              [0, 0, 2, 1],
-                              [0, 0, 1, 2],
-                              [0, 0, 2, 2],
-                              [0, 0, 1, 3],
-                              [0, 0, 2, 3]], dtype=np.float32)
-        raw_state[:, 0:2] = state[0][:]
-        
-    elif args.env == 'loothard':
-        raw_state = np.array([[0, 0, 0, 0],
-                              [0, 0, 1, 1],
-                              [0, 0, 2, 1],
-                              [0, 0, 1, 2],
-                              [0, 0, 2, 2],
-                              [0, 0, 3, 0],
-                              [0, 0, 0, 0]], dtype=np.float32)
-        raw_state[:, 0:2] = state[0][:]
-
-    elif args.env == 'loot':
-        raw_state = np.array([[0, 0, 0, 0],
-                              [0, 0, 1, 1],
-                              [0, 0, 2, 1],
-                              [0, 0, 1, 2],
-                              [0, 0, 2, 2],
-                              [0, 0, 0, 0],
-                              [0, 0, 0, 0]], dtype=np.float32)
-        raw_state[:, 0:2] = state[0][:]
-
-    elif args.env == 'lootcolor':
-        raw_state = np.array([[0, 0, 0, 0],
-                              [0, 0, 1, 10],
-                              [0, 0, 2, 10],
-                              [0, 0, 1, 20],
-                              [0, 0, 2, 20],
-                              [0, 0, 0, 0],
-                              [0, 0, 0, 0]], dtype=np.float32)
-        raw_state[:, 0:2] = state[0][:]
-
-    state = raw_state.reshape(-1)
-    state = state.tolist()
-    return torch.tensor(state).to(device)
-
-
 def simplify_action_loot(action):
     """simplify 9 actions to 5 actions
     """
